+++ extras/qpsclient.py
@@ -34,6 +34,10 @@ class QPSSpider(Spider):
         elif self.download_delay is not None:
             self.download_delay = float(self.download_delay)
 
+    async def yield_seeds(self):
+        for seed in self.start_requests():
+            yield seed
+
     def start_requests(self):
         url = self.benchurl
         if self.latency is not None: