+++ tests/CrawlerProcess/asyncio_enabled_reactor.py
@@ -38,8 +38,9 @@ class ReactorCheckExtension:
 class NoRequestsSpider(scrapy.Spider):
     name = "no_request"
 
-    def start_requests(self):
-        return []
+    async def yield_seeds(self):
+        return
+        yield
 
 
 process = CrawlerProcess(