+++ tests/CrawlerProcess/multi.py
@@ -5,8 +5,9 @@ from scrapy.crawler import CrawlerProcess
 class NoRequestsSpider(scrapy.Spider):
     name = "no_request"
 
-    def start_requests(self):
-        return []
+    async def yield_seeds(self):
+        return
+        yield
 
 
 process = CrawlerProcess(settings={})