+++ tests/CrawlerProcess/reactor_default.py
@@ -8,8 +8,9 @@ from scrapy.crawler import CrawlerProcess
 class NoRequestsSpider(scrapy.Spider):
     name = "no_request"
 
-    def start_requests(self):
-        return []
+    async def yield_seeds(self):
+        return
+        yield
 
 
 process = CrawlerProcess(settings={})