+++ tests/CrawlerProcess/reactor_select_subclass_twisted_reactor_select.py
@@ -17,8 +17,9 @@ installReactor(reactor)
 class NoRequestsSpider(scrapy.Spider):
     name = "no_request"
 
-    def start_requests(self):
-        return []
+    async def yield_seeds(self):
+        return
+        yield
 
 
 process = CrawlerProcess(