+++ tests/CrawlerProcess/reactor_select_twisted_reactor_select.py
@@ -9,8 +9,9 @@ selectreactor.install()
 class NoRequestsSpider(scrapy.Spider):
     name = "no_request"
 
-    def start_requests(self):
-        return []
+    async def yield_seeds(self):
+        return
+        yield
 
 
 process = CrawlerProcess(